Add a lock to syscalls that access the filesys.
This should have been implemented in project 2. This change passes
testcases that access the filesys in parallel eg. page-merge-mm. Not
entirely sure why, but sometimes in parallel access to the filesys, some
accesses see a blank root directory.

Open question is if we need to share the lock with load_page_from_filesys, but it seems like testcases are passing without it. Can consider a system-wide filesys lock if there are further issues.